gql test separation
